clientDashboard created and user model for client creation is changed

+++ backend/adsgrill_cms/micro_services/development/comments.py
@@ -19,45 +19,61 @@ class CsrfExemptSessionAuthentication(SessionAuthentication):
     
 class CommentsView(CsrfExemptMixin, APIView):
     authentication_classes = [CsrfExemptSessionAuthentication, SessionAuthentication]
-    # permission_classes = [IsAuthenticated]
+    permission_classes = [IsAuthenticated]
 
     def post(self, request):
         try:
             currentUser = request.user
             requestData = request.data
             issueID = requestData.get('issueID')
-            issueInstance = Issue.objects.get(pk=issueID)
-            assignee = issueInstance.assignee
-            sprintInstance = issueInstance.sprint
-            projectInstance = sprintInstance.project
-            if assignee!=currentUser:
+            sprintID = requestData.get('sprintID')
+            if issueID:
+                issueInstance = Issue.objects.get(pk=issueID)
+                assignee = issueInstance.assignee
+                sprintInstance = issueInstance.sprint
+                projectInstance = sprintInstance.project
+            if sprintID:
+                sprintInstance = Sprint.objects.get(pk=sprintID)
+                projectInstance = sprintInstance.project
+            if currentUser.role.name == 'client' or currentUser.role.name == 'Project Manager' or assignee==currentUser:
+                attachments = request.FILES.getlist('attachments', [])
+                attachment_file_names = []
+                if attachments:
+                    if issueID:
+                        subDirectory = os.path.join('media', 'uploads', 'Development', 'comments', str(f"{projectInstance.key}_{sprintInstance.key}_{issueInstance.key}"))
+                    if sprintID:
+                         subDirectory = os.path.join('media', 'uploads', 'Development', 'client', 'comments', str(f"{projectInstance.key}_{sprintInstance.key}"))
+                    if not os.path.exists(subDirectory):
+                        os.makedirs(subDirectory)
+                    for attachment in attachments:
+                        file_extension = attachment.name.split('.')[-1]
+                        unique_id = str(uuid.uuid4().hex[:6])
+                        if issueID:
+                            unique_filename = f"{unique_id}_{issueInstance.key}.{file_extension}"
+                        if sprintID:
+                            unique_filename = f"{unique_id}_{sprintInstance.key}.{file_extension}"
+                        attachment_path = os.path.join(subDirectory, unique_filename)
+
+                        with open(attachment_path, 'wb') as destination:
+                                for chunk in attachment.chunks():
+                                    destination.write(chunk)
+
+                        attachment_file_names.append(attachment_path)
+                with transaction.atomic():
+                    commentInstance = {
+                            'author': currentUser,
+                            'description': requestData.get('desc'),
+                        }
+                    if sprintID:
+                        commentInstance['sprint'] = sprintInstance
+                    if issueID:
+                        commentInstance['issue'] = issueInstance
+                    if attachment_file_names:
+                        commentInstance['attachment'] = attachment_file_names
+                    commentInstance = Comment.objects.create(**commentInstance)
+                    # commentInstance.save()
+            else:
                 return JsonResponse({"message":"You dont have access to comment on other's issues"})
-            
-            attachments = request.FILES.getlist('attachments', [])
-            attachment_file_names = []
-            subDirectory = os.path.join('media', 'uploads', 'Development', 'comments', str(f"{projectInstance.key}_{sprintInstance.key}_{issueInstance.key}"))
-            if not os.path.exists(subDirectory):
-                os.makedirs(subDirectory)
-            for attachment in attachments:
-                file_extension = attachment.name.split('.')[-1]
-                unique_id = str(uuid.uuid4().hex[:6])
-                unique_filename = f"{unique_id}_{issueInstance.key}.{file_extension}"
-                attachment_path = os.path.join(subDirectory, unique_filename)
-
-                with open(attachment_path, 'wb') as destination:
-                        for chunk in attachment.chunks():
-                            destination.write(chunk)
-
-                attachment_file_names.append(attachment_path)
-            with transaction.atomic():
-                commentInstance = Comment.objects.create(
-                        sprint = sprintInstance,
-                        issue = issueInstance,
-                        author = currentUser,
-                        description = requestData.get('desc'),
-                        attachment = attachment_file_names
-                    )
-                commentInstance.save()
         except IndentationError as i:
             return JsonResponse({"error":str(i)})
         except Exception as e:
@@ -101,14 +117,15 @@ class CommentsView(CsrfExemptMixin, APIView):
                     allComments = Comment.objects.filter(issue=issueInstance, author__role__name='development').order_by('-created_at')
             responseData = [{
                 "commentID": comment.pk,
-                "sprintID":comment.sprint.pk,
-                "IssueID":comment.issue.pk,
+                "sprintID":comment.sprint.pk if comment.sprint else None,
+                "IssueID":comment.issue.pk if comment.issue else None,
                 "author":{
                     "id":comment.author.pk,
                     "name":comment.author.name,
                 },
                 "description":comment.description,
-                "attachments":comment.attachment
+                "attachments":comment.attachment,
+                "created_at":comment.created_at
             }for comment in allComments]
 
         except Exception as e:
@@ -119,35 +136,43 @@ class CommentsView(CsrfExemptMixin, APIView):
         try:
             requestData = request.data
             currentUser = request.user
-            commentInstance = Comment.objects.get(pk=requestData.get('id'))
+            commentInstance = Comment.objects.get(pk=requestData.get('commentID'))
+            desc = requestData.get('desc')
             issueInstance = commentInstance.issue
             sprintInstance = commentInstance.sprint
             projectInstance = sprintInstance.project
 
-            if commentInstance.issue.assignee != currentUser:
-                return JsonResponse({"message":"You dont have access to make changes in comment on other's issues"})
-            attachments = request.FILES.getlist('attachments')
+            if currentUser.role.name == 'client' or currentUser.role.name == 'Project Manager' or commentInstance.issue.assignee == currentUser:
+                attachments = request.FILES.getlist('attachments')
+                attachment_file_names = []
+                if attachments:
+                    if issueInstance:
+                        subDirectory = os.path.join('media', 'uploads', 'Development', 'comments', str(f"{projectInstance.key}_{sprintInstance.key}_{issueInstance.key}"))
+                    if sprintInstance:
+                       subDirectory = os.path.join('media', 'uploads', 'Development', 'client', 'comments', str(f"{projectInstance.key}_{sprintInstance.key}")) 
+                    if not os.path.exists(subDirectory):
+                        os.makedirs(subDirectory)
+                    for attachment in attachments:
+                        file_extension = attachment.name.split('.')[-1]
+                        unique_id = str(uuid.uuid4().hex[:6])
+                        if issueInstance:
+                            unique_filename = f"{unique_id}_{issueInstance.key}.{file_extension}"
+                        if sprintInstance:
+                            unique_filename = f"{unique_id}_{sprintInstance.key}.{file_extension}"
+                        attachment_path = os.path.join(subDirectory, unique_filename)
 
-            subDirectory = os.path.join('media', 'uploads', 'Development', 'comments', str(f"{projectInstance.key}_{sprintInstance.key}_{issueInstance.key}"))
-            if not os.path.exists(subDirectory):
-                os.makedirs(subDirectory)
-            attachment_file_names = []
-            for attachment in attachments:
-                file_extension = attachment.name.split('.')[-1]
-                unique_id = str(uuid.uuid4().hex[:6])
-                unique_filename = f"{unique_id}_{issueInstance.key}.{file_extension}"
-                attachment_path = os.path.join(subDirectory, unique_filename)
+                        with open(attachment_path, 'wb') as destination:
+                                for chunk in attachment.chunks():
+                                    destination.write(chunk)
 
-                with open(attachment_path, 'wb') as destination:
-                        for chunk in attachment.chunks():
-                            destination.write(chunk)
-
-                attachment_file_names.append(attachment_path)
-            with transaction.atomic():
-                commentInstance.description = requestData.get('desc'),
-                commentInstance.attachment.extend(attachment_file_names)
-                commentInstance.save()
-
+                        attachment_file_names.append(attachment_path)
+                with transaction.atomic():
+                    commentInstance.description = desc
+                    if attachment_file_names:
+                        commentInstance.attachment.extend(attachment_file_names)
+                    commentInstance.save()
+            else:
+                return JsonResponse({'message':"You dont have access to make changes in comment on other's issues"})
         except IntegrityError as i:
             return JsonResponse({"message":str(i)})
         except Exception as e:
@@ -157,7 +182,7 @@ class CommentsView(CsrfExemptMixin, APIView):
 
     def delete(self, request):
         try:
-            commentInstance = Comment.objects.get(pk=request.GET.get('id'))
+            commentInstance = Comment.objects.get(pk=request.GET.get('commentID'))
             issueInstance = commentInstance.issue
             sprintInstance = commentInstance.sprint
             projectInstance = sprintInstance.project
@@ -167,10 +192,16 @@ class CommentsView(CsrfExemptMixin, APIView):
                 for file_path in file_paths:
                     if os.path.exists(file_path):
                         os.remove(file_path)
-
-            directory_path = os.path.join('media', 'uploads', 'Development', 'comments', str(f"{projectInstance.key}_{sprintInstance.key}_{issueInstance.key}"))
+            if issueInstance:
+                directory_path = os.path.join('media', 'uploads', 'Development', 'comments', str(f"{projectInstance.key}_{sprintInstance.key}_{issueInstance.key}"))
+            if sprintInstance:
+                directory_path = os.path.join('media', 'uploads', 'Development', 'client', 'comments', str(f"{projectInstance.key}_{sprintInstance.key}"))
+            print('directory-path', directory_path)
             if os.path.exists(directory_path):
-                os.rmdir(directory_path)
+                if not os.listdir(directory_path):
+                    os.rmdir(directory_path)
+                else:
+                    pass
             
             commentInstance.delete()
 
@@ -187,8 +218,9 @@ class DownloadCommentsAttachments(CsrfExemptMixin, APIView):
 
     def get(self, request):
         try:
-            commentInstance = Comment.objects.get(pk=request.GET.get('id'))
+            commentInstance = Comment.objects.get(pk=request.GET.get('commentID'))
             issueInstance = commentInstance.issue
+            sprintInstance = commentInstance.sprint
             files = commentInstance.attachment
 
             if files:
@@ -198,7 +230,10 @@ class DownloadCommentsAttachments(CsrfExemptMixin, APIView):
                         filename = file.split("\\")[-1]
                         commentZip.write(file, filename)
                 response = HttpResponse(zip_buffer.getvalue(), content_type = 'application/zip')
-                response['Content-Disposition'] = f'attachment; filename=comments_{issueInstance.key}_attachments.zip'
+                if issueInstance:
+                    response['Content-Disposition'] = f'attachment; filename=comments_{issueInstance.key}_attachments.zip'
+                if sprintInstance:
+                    response['Content-Disposition'] = f'attachment; filename=comments_{sprintInstance.key}_attachments.zip'
 
             else:
                 response = JsonResponse({"message":"No files found for the specified criteria"}, status=status.HTTP_204_NO_CONTENT)
